backend/requirements.txt
backports.functools-lru-cache==1.6.6
bcrypt==3.1.7
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
click==7.1.2
configparser==4.0.2
contextlib2==0.6.0.post1
cycler==0.10.0
Flask==1.1.4
idna==2.10
importlib-metadata==2.1.3
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==2.2.5
numpy==1.16.6
pathlib2==2.3.7.post1
pycparser==2.21
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-dotenv==0.18.0
pytz==2024.1
requests==2.27.1
scandir==1.10.0
six==1.16.0
SQLAlchemy==1.4.54
subprocess32==3.5.4
twilio==6.63.2
typing==3.10.0.0
urllib3==1.26.20
Werkzeug==1.0.1
zipp==1.2.0
